+++ app/Laravel/Builder.php
@@ -7,14 +7,16 @@ use Illuminate\Container\Container;
 use Illuminate\Database\Eloquent\Builder as BaseBuilder;
 use Illuminate\Pagination\Paginator;
 
-class Builder extends BaseBuilder {
-	public function simplePaginate($perPage = null, $columns = ['*'], $pageName = 'page', $currentPage = null) {
-		$currentPage = $currentPage ? : Paginator::resolveCurrentPage($pageName);
+class Builder extends BaseBuilder
+{
+	public function simplePaginate($perPage = null, $columns = ['*'], $pageName = 'page', $page = null)
+	{
+		$page = $page ?: Paginator::resolveCurrentPage($pageName);
 
-		$perPage = $perPage ? : $this->model->getPerPage();
+		$perPage = $perPage ?: $this->model->getPerPage();
 
 		$items = ($total = $this->toBase()->getCountForPagination())
-			? $this->forPage($currentPage, $perPage)->get($columns)
+			? $this->forPage($page, $perPage)->get($columns)
 			: $this->model->newCollection();
 
 		$options = [
@@ -23,11 +25,12 @@ class Builder extends BaseBuilder {
 		];
 
 		return Container::getInstance()->makeWith(SimplePaginator::class, compact(
-			'items', 'total', 'perPage', 'currentPage', 'options'
+			'items', 'total', 'perPage', 'page', 'options'
 		));
 	}
 
-	protected function simplePaginator($items, $perPage, $currentPage, $options) {
+	protected function simplePaginator($items, $perPage, $currentPage, $options)
+	{
 		return Container::getInstance()->makeWith(SimplePaginator::class, compact(
 			'items', 'perPage', 'currentPage', 'options'
 		));