+++ CHANGELOG.md
@@ -4,11 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.0.3 - 2021-08-12]
+## [2.0.4] - 2022-09-23
+### Added
+- Tooltip for Towers.
+- Defense Units in Tower Tooltip.
+
+### Changed
+- Updated all defense units for old towers.
+
+## [2.0.3] - 2021-08-12
 ### Fixed
 - Pagination do not used the selected filters. 
 
-## [2.0.2 - 2021-08-12]
+## [2.0.2] - 2021-08-12
 ### Added
 - Multiple Filter for Game Modes and Difficulties.
 ### Fixed